requirements/local.txt
# Local development dependencies go here
-r base.txt
ipython==1.1.0
django_extensions==1.2.5
nose==1.3.0
django-nose==1.2
nose-notify==0.4.2
Sphinx==1.2b1
